[network] move network_builder into its own crate and setup_network out of main_node
