+++ libra-node/src/main_node.rs
@@ -8,8 +8,7 @@ use executor::{db_bootstrapper::bootstrap_db_if_empty, Executor};
 use executor_types::ChunkExecutor;
 use futures::{channel::mpsc::channel, executor::block_on};
 use libra_config::{
-    chain_id::ChainId,
-    config::{DiscoveryMethod, NetworkConfig, NodeConfig, RoleType},
+    config::{NetworkConfig, NodeConfig, RoleType},
     utils::get_genesis_txn,
 };
 use libra_json_rpc::bootstrap_from_config as bootstrap_rpc;
@@ -17,19 +16,16 @@ use libra_logger::prelude::*;
 use libra_mempool::gen_mempool_reconfig_subscription;
 use libra_metrics::metric_server;
 use libra_secure_storage::config;
-use libra_types::waypoint::Waypoint;
 use libra_vm::LibraVM;
 use libradb::LibraDB;
-use network::validator_network::network_builder::{AuthenticationMode, NetworkBuilder};
 use network_simple_onchain_discovery::{
     gen_simple_discovery_reconfig_subscription, ConfigurationChangeListener,
 };
-use onchain_discovery::builder::OnchainDiscoveryBuilder;
 use state_synchronizer::StateSynchronizer;
-use std::{boxed::Box, collections::HashMap, net::ToSocketAddrs, sync::Arc, thread, time::Instant};
-use storage_interface::{DbReader, DbReaderWriter};
+use std::{boxed::Box, net::ToSocketAddrs, sync::Arc, thread, time::Instant};
+use storage_interface::DbReaderWriter;
 use storage_service::start_storage_service_with_db;
-use tokio::runtime::{Builder, Runtime};
+use tokio::runtime::Runtime;
 
 const AC_SMP_CHANNEL_BUFFER_SIZE: usize = 1_024;
 const INTRA_NODE_CHANNEL_BUFFER_SIZE: usize = 1;
@@ -59,100 +55,6 @@ fn setup_debug_interface(config: &NodeConfig) -> NodeDebugService {
     .unwrap();
 
     NodeDebugService::new(addr)
-}
-
-// TODO(abhayb): Move to network crate (similar to consensus).
-pub fn setup_network(
-    chain_id: &ChainId,
-    role: RoleType,
-    config: &mut NetworkConfig,
-    libra_db: Arc<dyn DbReader>,
-    waypoint: Waypoint,
-) -> (Runtime, NetworkBuilder) {
-    let runtime = Builder::new()
-        .thread_name("network-")
-        .threaded_scheduler()
-        .enable_all()
-        .build()
-        .expect("Failed to start runtime. Won't be able to start networking.");
-
-    let identity_key = config::identity_key(config);
-    let peer_id = config::peer_id(config);
-
-    let mut network_builder = NetworkBuilder::new(
-        runtime.handle().clone(),
-        chain_id.clone(),
-        config.network_id.clone(),
-        role,
-        peer_id,
-        config.listen_address.clone(),
-    );
-    network_builder.add_connection_monitoring();
-
-    // Sanity check seed peer addresses.
-    config
-        .verify_seed_peer_addrs()
-        .expect("Seed peer addresses must be well-formed");
-    let seed_peers = config.seed_peers.clone();
-
-    if config.mutual_authentication {
-        let network_peers = config.network_peers.clone();
-        let trusted_peers = if role == RoleType::Validator {
-            // for validators, trusted_peers is empty will be populated from consensus
-            HashMap::new()
-        } else {
-            network_peers
-        };
-
-        info!(
-            "network setup: role: {}, seed_peers: {:?}, trusted_peers: {:?}",
-            role, seed_peers, trusted_peers,
-        );
-
-        network_builder
-            .authentication_mode(AuthenticationMode::Mutual(identity_key))
-            .trusted_peers(trusted_peers)
-            .seed_peers(seed_peers)
-            .connectivity_check_interval_ms(config.connectivity_check_interval_ms)
-            .add_connectivity_manager();
-    } else {
-        // Enforce the outgoing connection (dialer) verifies the identity of the listener (server)
-        network_builder.authentication_mode(AuthenticationMode::ServerOnly(identity_key));
-        if !seed_peers.is_empty() {
-            network_builder
-                .seed_peers(seed_peers)
-                .add_connectivity_manager();
-        }
-    }
-
-    match &config.discovery_method {
-        DiscoveryMethod::Gossip(gossip_config) => {
-            network_builder
-                .advertised_address(gossip_config.advertised_address.clone())
-                .discovery_interval_ms(gossip_config.discovery_interval_ms)
-                .add_gossip_discovery();
-        }
-        DiscoveryMethod::Onchain => {
-            let (network_tx, discovery_events) = network_builder.add_protocol_handler(
-                onchain_discovery::network_interface::network_endpoint_config(),
-            );
-            let onchain_discovery_builder = OnchainDiscoveryBuilder::build(
-                network_builder
-                    .conn_mgr_reqs_tx()
-                    .expect("ConnectivityManager must be installed"),
-                network_tx,
-                discovery_events,
-                network_builder.network_context().clone(),
-                libra_db,
-                waypoint,
-                runtime.handle(),
-            );
-            onchain_discovery_builder.start(runtime.handle());
-        }
-        DiscoveryMethod::None => {}
-    }
-
-    (runtime, network_builder)
 }
 
 pub fn setup_environment(node_config: &mut NodeConfig) -> LibraHandle {
@@ -225,7 +127,7 @@ pub fn setup_environment(node_config: &mut NodeConfig) -> LibraHandle {
     // Instantiate every network and collect the requisite endpoints for state_sync, mempool, and consensus.
     for (role, network_config) in network_configs {
         // Perform common instantiation steps
-        let (runtime, mut network_builder) = setup_network(
+        let (runtime, mut network_builder) = network_builder::builder::setup_network(
             &node_config.base.chain_id,
             role,
             network_config,
