+++ network/Cargo.toml
@@ -15,9 +15,11 @@ bytes = { version = "0.5.4", features = ["serde"] }
 futures = "0.3.5"
 hex = "0.4.2"
 once_cell = "1.4.0"
-pin-project = "0.4.22"
+pin-project = "0.4.20"
+proptest = { version = "0.10.0", default-features = true, optional = true }
 rand = "0.7.3"
-serde = { version = "1.0.112", default-features = false }
+rand_core = { version = "0.5.1", optional = true }
+serde = { version = "1.0.111", default-features = false }
 serde_bytes = "0.11.5"
 thiserror = "1.0.20"
 tokio = { version = "0.2.21", features = ["full"] }
@@ -32,22 +34,22 @@ libra-crypto = { path = "../crypto/crypto", version = "0.1.0" }
 libra-crypto-derive = { path = "../crypto/crypto-derive", version = "0.1.0" }
 libra-logger = { path = "../common/logger", version = "0.1.0" }
 libra-metrics = { path = "../common/metrics", version = "0.1.0" }
-libra-network-address = { path = "network-address", version = "0.1.0" }
+libra-network-address = { path = "../network/network-address", version = "0.1.0" }
 libra-proptest-helpers = { path = "../common/proptest-helpers", version = "0.1.0", optional = true }
 libra-security-logger = { path = "../common/security-logger", version = "0.1.0" }
 libra-types = { path = "../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../common/workspace-hack", version = "0.1.0" }
-memsocket = { path = "memsocket", version = "0.1.0" }
-netcore = { path = "netcore", version = "0.1.0" }
+memsocket = { path = "../network/memsocket", version = "0.1.0" }
+netcore = { path = "../network/netcore", version = "0.1.0" }
 num-variants = { path = "../common/num-variants", version = "0.1.0" }
-proptest = { version = "0.10.0", default-features = true, optional = true }
 stream-ratelimiter = { path = "../common/stream-ratelimiter", version = "0.1.0" }
-rand_core = { version = "0.5.1", optional = true }
 
 [dev-dependencies]
 criterion = "0.3.2"
 serial_test = "0.4.0"
-socket-bench-server = { path = "socket-bench-server", version = "0.1.0" }
+
+network-builder = {path = "../network/builder", version = "0.1.0"}
+socket-bench-server = { path = "../network/socket-bench-server", version = "0.1.0" }
 
 [features]
 default = []
