+++ network/builder/Cargo.toml
@@ -13,17 +13,23 @@ edition = "2018"
 
 [dependencies]
 futures = "0.3.5"
+rand = "0.7.3"
+serde = { version = "1.0.111", default-features = false }
 tokio = { version = "0.2.21", features = ["full"] }
 tokio-retry = "0.2.0"
 
 
 channel = { path = "../../common/channel", version = "0.1.0" }
-network = { path = "../", version = "0.1.0" }
+lcs = { path = "../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-config = { path = "../../config", version = "0.1.0" }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
 libra-logger = { path = "../../common/logger", version = "0.1.0" }
 libra-metrics = {path = "../../common/metrics", version = "0.1.0"}
-libra-network-address = { path = "../network-address", version = "0.1.0" }
+libra-network-address = { path = "../../network/network-address", version = "0.1.0" }
+libra-secure-storage = { path = "../../secure/storage", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 netcore = { path = "../../network/netcore", version = "0.1.0" }
+network = { path = "../../network", version = "0.1.0" }
+onchain-discovery = { path = "../../network/onchain-discovery", version = "0.1.0" }
+storage-interface= { path = "../../storage/storage-interface", version = "0.1.0" }
