+++ network/src/lib.rs
@@ -16,12 +16,11 @@ pub mod interface;
 pub mod logging;
 pub mod peer_manager;
 pub mod protocols;
-pub mod validator_network;
 
 pub mod counters;
 mod peer;
 mod sink;
-mod transport;
+pub mod transport;
 
 #[cfg(not(any(feature = "testing", feature = "fuzzing")))]
 mod noise;
