--- network/src/validator_network/mod.rs
@@ -1,7 +1,0 @@
-// Copyright (c) The Libra Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-//! Network API for Libra
-
-pub use crate::protocols::rpc::error::RpcError;
-pub mod network_builder;
