+++ state-synchronizer/src/tests/integration_tests.rs
@@ -23,7 +23,7 @@ use libra_types::{
     validator_info::ValidatorInfo, validator_signer::ValidatorSigner,
     validator_verifier::random_validator_verifier, waypoint::Waypoint,
 };
-use network::validator_network::network_builder::{AuthenticationMode, NetworkBuilder};
+use network_builder::builder::{AuthenticationMode, NetworkBuilder};
 use rand::{rngs::StdRng, SeedableRng};
 use std::{
     collections::HashMap,
